Remove unnecessary implementations of CleaningJavaCompiler
