+++ subprojects/language-java/src/main/java/org/gradle/api/internal/tasks/compile/CleaningJavaCompiler.java
@@ -16,29 +16,50 @@
 package org.gradle.api.internal.tasks.compile;
 
 import org.gradle.api.internal.TaskOutputsInternal;
+import org.gradle.api.tasks.WorkResult;
 import org.gradle.internal.file.Deleter;
 import org.gradle.language.base.internal.compile.Compiler;
 import org.gradle.language.base.internal.tasks.SimpleStaleClassCleaner;
 import org.gradle.language.base.internal.tasks.StaleClassCleaner;
 
-public class CleaningJavaCompiler extends CleaningJavaCompilerSupport<JavaCompileSpec> implements org.gradle.language.base.internal.compile.Compiler<JavaCompileSpec> {
-    private final Compiler<JavaCompileSpec> compiler;
+import javax.annotation.Nullable;
+import java.io.File;
+
+/**
+ * Deletes stale classes before invoking the actual compiler.
+ */
+public class CleaningJavaCompiler<T extends JavaCompileSpec> implements Compiler<T> {
+    private final Compiler<T> compiler;
     private final TaskOutputsInternal taskOutputs;
     private final Deleter deleter;
 
-    public CleaningJavaCompiler(Compiler<JavaCompileSpec> compiler, TaskOutputsInternal taskOutputs, Deleter deleter) {
+    public CleaningJavaCompiler(Compiler<T> compiler, TaskOutputsInternal taskOutputs, Deleter deleter) {
         this.compiler = compiler;
         this.taskOutputs = taskOutputs;
         this.deleter = deleter;
     }
 
     @Override
-    public Compiler<JavaCompileSpec> getCompiler() {
-        return compiler;
+    public WorkResult execute(T spec) {
+        StaleClassCleaner cleaner = new SimpleStaleClassCleaner(deleter, taskOutputs);
+
+        addDirectory(cleaner, spec.getDestinationDir());
+        MinimalJavaCompileOptions compileOptions = spec.getCompileOptions();
+        addDirectory(cleaner, compileOptions.getAnnotationProcessorGeneratedSourcesDirectory());
+        addDirectory(cleaner, compileOptions.getHeaderOutputDirectory());
+        cleaner.execute();
+
+        Compiler<? super T> compiler = getCompiler();
+        return compiler.execute(spec);
     }
 
-    @Override
-    protected StaleClassCleaner createCleaner(final JavaCompileSpec spec) {
-        return new SimpleStaleClassCleaner(deleter, taskOutputs);
+    private void addDirectory(@Nullable StaleClassCleaner cleaner, File dir) {
+        if (dir != null) {
+            cleaner.addDirToClean(dir);
+        }
+    }
+
+    public Compiler<T> getCompiler() {
+        return compiler;
     }
 }
